chore: rename entry/xxx to entry-xxx
