+++ packages/react-server/tsup.config.ts
@@ -10,9 +10,9 @@ export default defineConfig([
       "src/runtime-browser.ts",
       "src/runtime-ssr.ts",
       "src/runtime-server.ts",
-      "src/entry/server.tsx",
-      "src/entry/ssr.tsx",
-      "src/entry/browser.tsx",
+      "src/entry-server.tsx",
+      "src/entry-ssr.tsx",
+      "src/entry-browser.tsx",
       "src/plugin/index.ts",
     ],
     format: ["esm"],
